+++ crates/core/action-context/src/lib.rs
@@ -33,8 +33,6 @@ pub struct ActionContext {
 
     pub initial_targets: FxHashSet<TargetLocator>,
 
-    pub interactive: bool,
-
     pub passthrough_args: Vec<String>,
 
     pub primary_targets: FxHashSet<Target>,