+++ crates/core/runner/src/runner.rs
@@ -3,7 +3,7 @@ use crate::target_hash::TargetHasher;
 use crate::{errors::RunnerError, inputs_collector};
 use console::Term;
 use miette::IntoDiagnostic;
-use moon_action::{ActionStatus, Attempt};
+use moon_action::{ActionNode, ActionStatus, Attempt};
 use moon_action_context::{ActionContext, TargetState};
 use moon_cache_item::CacheItem;
 use moon_config::{TaskOptionAffectedFiles, TaskOutputStyle};
@@ -38,6 +38,8 @@ pub enum HydrateFrom {
 pub struct Runner<'a> {
     pub cache: CacheItem<RunTargetState>,
 
+    pub node: Option<ActionNode>,
+
     emitter: &'a Emitter,
 
     project: &'a Project,
@@ -68,6 +70,7 @@ impl<'a> Runner<'a> {
 
         Ok(Runner {
             cache,
+            node: None,
             emitter,
             project,
             stderr: Term::buffered_stderr(),
@@ -542,14 +545,15 @@ impl<'a> Runner<'a> {
         let primary_longest_width = context.primary_targets.iter().map(|t| t.id.len()).max();
         let is_primary = context.primary_targets.contains(&self.task.target);
         let is_real_ci = is_ci() && !is_test_env();
-        let is_persistent = self.task.is_persistent();
+        let is_persistent =
+            self.node.as_ref().is_some_and(|n| n.is_persistent()) || self.task.is_persistent();
         let output;
         let error;
 
         // When a task is configured as local (no caching), or the interactive flag is passed,
         // we don't "capture" stdout/stderr (which breaks stdin) and let it stream natively.
         let is_interactive = (!self.task.options.cache && context.primary_targets.len() == 1)
-            || context.interactive
+            || self.node.as_ref().is_some_and(|n| n.is_interactive())
             || self.task.is_interactive();
 
         // When the primary target, always stream the output for a better developer experience.